Create IR and Wi-Fi
This code uses a Scanner to read an option from the user, either 1 for IR or 2 for Wi-Fi. The appropriate method is then called based on the selected option. In the sendIRCommand method, the code to send an IR command would go. In the sendWifiCommand method, a DatagramSocket is used to send a command as a UDP packet to a specific IP address and port.